refactor(approval-gate): extract state.rs from lib.rs
Third step of breaking lib.rs into focused modules. This commit lifts
the iii-backed state-bus and function-executor implementations into
their own module, along with the `__from_approval` marker plumbing
and the boot-time marker-target safety check.

Moved (lib.rs -> state.rs):
- StateBus trait + IiiStateBus impl  (iii state::* wrapper)
- FunctionExecutor trait + IiiFunctionExecutor impl  (iii.trigger wrapper)
- rule_for()  (lookup helper used by IiiFunctionExecutor)
- merge_from_approval_marker_if_needed()  (marker payload composer)
- unverified_marker_targets()  (boot guard)

The StateBus / FunctionExecutor traits are kept exactly as they were
in lib.rs — they exist purely as test seams so unit tests can swap in
InMemoryStateBus / FakeExecutor. No new abstractions added.

Public items stay re-exported from the crate root.
Pure move, no behavior change. 141 approval-gate tests pass;
all 8 downstream worker crates pass (turn-orchestrator's pre-existing
dual_write flake notwithstanding).
