+++ approval-gate/src/state.rs
@@ -1,0 +1,179 @@
+//! State-store and function-executor traits, plus their iii-backed
+//! implementations and the `__from_approval` marker plumbing.
+//!
+//! The traits exist purely as test seams — unit tests swap in
+//! `InMemoryStateBus` / `FakeExecutor` while production code uses the
+//! `Iii*` implementations that call iii directly. No new abstractions
+//! beyond what's needed for that seam.
+
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use iii_sdk::{IIIError, TriggerRequest, III};
+use serde_json::{json, Value};
+
+use crate::config::InterceptorRule;
+
+/// Look up the [`InterceptorRule`] for `function_id`, if one is configured.
+/// Pure helper; no I/O. Used by the gate's intercept flow and by the
+/// production [`IiiFunctionExecutor`] to decide whether to inject the
+/// `__from_approval` marker.
+pub(crate) fn rule_for<'a>(
+    rules: &'a [InterceptorRule],
+    function_id: &str,
+) -> Option<&'a InterceptorRule> {
+    rules.iter().find(|r| r.function_id == function_id)
+}
+
+/// Stamp the `__from_approval` marker onto a function call's args when the
+/// rule asks for it. The marker carries `{ call_id, session_id }` so the
+/// target function can validate the call came through approval-gate (via
+/// `approval::lookup_record`) instead of via direct trigger bypass.
+///
+/// Idempotent on shape: object args get the marker merged in; null args
+/// become `{ __from_approval: ... }`; any other shape (array, scalar)
+/// gets wrapped as `{ payload, __from_approval: ... }` so it stays
+/// recoverable on the target side.
+pub(crate) fn merge_from_approval_marker_if_needed(
+    inject: bool,
+    args: Value,
+    function_call_id: &str,
+    session_id: &str,
+) -> Value {
+    if !inject {
+        return args;
+    }
+    let marker = json!({
+        "call_id": function_call_id,
+        "session_id": session_id,
+    });
+    match args {
+        Value::Object(mut m) => {
+            m.insert("__from_approval".into(), marker);
+            Value::Object(m)
+        }
+        other if other.is_null() => json!({ "__from_approval": marker }),
+        other => json!({
+            "payload": other,
+            "__from_approval": marker,
+        }),
+    }
+}
+
+/// Abstraction over the iii state bus — the kv layer where pending and
+/// resolved approval records live. Exists so unit tests can swap in a
+/// `BTreeMap`-backed fake; production uses [`IiiStateBus`].
+#[async_trait]
+pub trait StateBus: Send + Sync {
+    async fn set(&self, scope: &str, key: &str, value: Value) -> Result<(), IIIError>;
+    async fn get(&self, scope: &str, key: &str) -> Option<Value>;
+    async fn list_prefix(&self, scope: &str, prefix: &str) -> Vec<Value>;
+}
+
+/// Invokes an iii function with arguments and returns its result or an
+/// error string. Abstracted so tests can stub the underlying call.
+#[async_trait]
+pub trait FunctionExecutor: Send + Sync {
+    async fn invoke(
+        &self,
+        function_id: &str,
+        args: Value,
+        function_call_id: &str,
+        session_id: &str,
+    ) -> Result<Value, String>;
+}
+
+/// Production [`FunctionExecutor`] backed by `iii.trigger`.
+pub struct IiiFunctionExecutor {
+    pub iii: III,
+    pub rules: Arc<Vec<InterceptorRule>>,
+}
+
+#[async_trait]
+impl FunctionExecutor for IiiFunctionExecutor {
+    async fn invoke(
+        &self,
+        function_id: &str,
+        args: Value,
+        function_call_id: &str,
+        session_id: &str,
+    ) -> Result<Value, String> {
+        let inject =
+            rule_for(self.rules.as_slice(), function_id).is_some_and(|r| r.inject_approval_marker);
+        let payload =
+            merge_from_approval_marker_if_needed(inject, args, function_call_id, session_id);
+        self.iii
+            .trigger(TriggerRequest {
+                function_id: function_id.to_string(),
+                payload,
+                action: None,
+                timeout_ms: None,
+            })
+            .await
+            .map_err(|e| e.to_string())
+    }
+}
+
+/// Production [`StateBus`] backed by iii's `state::*` builtins.
+pub struct IiiStateBus(pub III);
+
+#[async_trait]
+impl StateBus for IiiStateBus {
+    async fn set(&self, scope: &str, key: &str, value: Value) -> Result<(), IIIError> {
+        self.0
+            .trigger(TriggerRequest {
+                function_id: "state::set".into(),
+                payload: json!({ "scope": scope, "key": key, "value": value }),
+                action: None,
+                timeout_ms: None,
+            })
+            .await
+            .map(|_| ())
+    }
+    async fn get(&self, scope: &str, key: &str) -> Option<Value> {
+        self.0
+            .trigger(TriggerRequest {
+                function_id: "state::get".into(),
+                payload: json!({ "scope": scope, "key": key }),
+                action: None,
+                timeout_ms: None,
+            })
+            .await
+            .ok()
+            .filter(|v| !v.is_null())
+    }
+    async fn list_prefix(&self, scope: &str, prefix: &str) -> Vec<Value> {
+        let resp = self
+            .0
+            .trigger(TriggerRequest {
+                function_id: "state::list".into(),
+                payload: json!({ "scope": scope, "prefix": prefix }),
+                action: None,
+                timeout_ms: None,
+            })
+            .await
+            .unwrap_or_else(|_| json!({ "items": [] }));
+        // Engine may return either {"items": [...]} or a plain Array.
+        if let Some(arr) = resp.as_array() {
+            return arr.clone();
+        }
+        resp.get("items")
+            .and_then(|v| v.as_array().cloned())
+            .unwrap_or_default()
+            .into_iter()
+            .map(|entry| entry.get("value").cloned().unwrap_or(entry))
+            .collect()
+    }
+}
+
+/// Return the list of function ids whose interceptor asks the gate to
+/// inject `__from_approval` without asserting that the target validates it.
+/// Empty list ⇒ config is safe to register. Pure — exposed for tests and
+/// for the boot-time check in `register`.
+pub fn unverified_marker_targets(rules: &[InterceptorRule]) -> Vec<&str> {
+    rules
+        .iter()
+        .filter(|r| r.inject_approval_marker && !r.marker_target_verified)
+        .map(|r| r.function_id.as_str())
+        .collect()
+}
